[DROOLS-7317] unify SessionConfiguation and SessionConfiguationImpl
