+++ drools-model/drools-model-codegen/src/main/resources/class-templates/rules/ProjectRuntimeJavaTemplate.java
@@ -69,7 +69,7 @@ public class ProjectRuntime implements KieRuntimeBuilder {
     }
 
     private org.kie.api.runtime.KieSessionConfiguration getConfForSession(String sessionName) {
-        org.drools.core.SessionConfigurationImpl conf = new org.drools.core.SessionConfigurationImpl();
+        org.drools.core.SessionConfiguration conf = new org.drools.core.SessionConfiguration();
         switch(sessionName) {
             // populated via codegen
         }
