+++ drools-persistence/drools-persistence-jpa/src/main/java/org/drools/persistence/PersistableRunner.java
@@ -92,7 +92,7 @@ public class PersistableRunner implements SingleSessionCommandService {
                               KieSessionConfiguration conf,
                               Environment env ) {
         if ( conf == null ) {
-            conf = SessionConfiguration.newInstance();
+            conf = new SessionConfiguration();
         }
         this.env = env;
 
@@ -168,7 +168,7 @@ public class PersistableRunner implements SingleSessionCommandService {
                               KieSessionConfiguration conf,
                               Environment env ) {
         if ( conf == null ) {
-            conf = SessionConfiguration.newInstance();
+            conf = new SessionConfiguration();
         }
 
         this.env = env;
