+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/mvel/integrationtests/OutOfMemoryTest.java
@@ -34,6 +34,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.kie.api.KieBase;
 import org.kie.api.runtime.KieSession;
+import org.kie.api.runtime.conf.KeepReferenceOption;
 import org.kie.api.runtime.rule.FactHandle;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -68,8 +69,8 @@ public class OutOfMemoryTest {
 
         int i = 0;
 
-        SessionConfiguration conf = SessionConfiguration.newInstance();
-        conf.setKeepReference( true ); // this is just for documentation purposes, since the default value is "true"
+        SessionConfiguration conf = new SessionConfiguration();
+        conf.setOption(KeepReferenceOption.YES ); // this is just for documentation purposes, since the default value is "true"
         try {
             for ( i = 0; i < 300000; i++ ) {
                 KieSession ksession = kbase.newKieSession( conf, null );
